docs/9.music/_archive/requirements_lyria.txt
# Requirements for Lyria stem generation script

google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0

